Adjusted ReadMe + Refactoring of UIButton and exctraction of method into UIImage extension
