+++ README.md
@@ -4,6 +4,6 @@ This package provides extension to commonly used UIKit Classes, like UIColor and
 This package is restricted to the iOS platfrom.
 
 
-This package is created by apploft on 18.12.18.
+This package is created by apploft on 02.04.2020.
 Copyright © 2019 apploft GmbH￼￼
 MIT License · http://choosealicense.com/licenses/mit/
